README.md: Consolidate WSL specific instructions
Move all of the documentation specific to the WSL kernel into README.md.
The README.md file is displayed in the GitHub web interface in favor of
the upstream README file.

Leave a symlink from README-Microsoft.WSL2 to README.md so that anyone
watching that file for changes will continue to see updates.

While consolidating documentation, clean up trailing spaces and adjust
the build dependency installation instructions to include the missing
apt command.

Signed-off-by: Tyler Hicks <tyhicks@linux.microsoft.com>
Signed-off-by: Kelsey Steele <kelseysteele@microsoft.com>