configs: build in select netfilter match/target modules
Certain netfilter module names are used twice, differing only in
capitalization. This poses a problem when copying the modules to Windows
environments that are not case-sensitive, as only one of the modules
will be copied.

Build in the necessary netfilter match/target modules until WSL can
provide all modules without risk of losing any.

Signed-off-by: Kelsey Steele <kelseysteele@microsoft.com>